requirement.txt
ultralytics
python-multipart
fastapi
uvicorn
ffmpegcv
opencv-python
httpx
ffmpeg-python
